Refactoring, seems to still work
