+++ social/app/models/post.py
@@ -4,9 +4,11 @@ import uuid
 import CommonMark
 from django.db import models
 from django.urls import reverse
+from django.db.models import Q
 
 from social.app.models.author import Author
 from social.app.models.category import Category
+from social.app.models.node import Node
 
 
 class Post(models.Model):
@@ -79,6 +81,7 @@ class Post(models.Model):
 
     unlisted = models.BooleanField(default=False)
 
+
     def get_absolute_url(self):
         return reverse('app:posts:detail', kwargs={'pk': self.id})
 
@@ -120,3 +123,65 @@ def keys(tuple_list):
     Accepts a tuple list, returns a list of each tuple's first values
     """
     return [x[0] for x in tuple_list]
+
+
+def get_all_public_posts():
+    public_posts = dict()
+    public_posts["user_posts"] = Post.objects \
+        .filter(Q(visibility="PUBLIC") | Q(visibility="SERVERONLY")) \
+        .order_by('-published')
+    return public_posts["user_posts"]
+
+
+def get_all_friend_posts(user):
+    friend_posts = dict()
+    friend_posts["user_posts"] = Post.objects \
+        .filter(author__id__in=user.friends.all()) \
+        .filter(Q(visibility="FRIENDS") | Q(visibility="PUBLIC") | Q(visibility="SERVERONLY")) \
+        .order_by('-published')
+    return friend_posts["user_posts"]
+
+
+def get_all_foaf_posts(author):
+    friends_list = set(f.id for f in author.friends.all())
+    foafs = set()
+
+    for friend in friends_list:
+        friend_obj = Author.objects.get(pk=friend)
+        new_foafs = set(ff.id for ff in friend_obj.friends.all())
+        foafs.update(new_foafs)
+    foafs.update(friends_list)
+
+    foaf_posts = dict()
+    foaf_posts["user_posts"] = Post.objects \
+        .filter(Q(author__id__in=foafs)) \
+        .filter(Q(visibility="FOAF") | Q(visibility="PUBLIC")).order_by('-published')
+
+    return foaf_posts["user_posts"]
+
+
+def get_remote_node_posts():
+    node_posts = list()
+    for node in Node.objects.filter(local=False):
+        for post_json in node.get_author_posts()['posts']:
+            author_json = post_json['author']
+            author = Author(
+                id=Author.get_id_from_uri(author_json['id']),
+                node=node,
+                displayName=author_json['displayName'],
+            )
+            post = Post(
+                id=post_json['id'],
+                title=post_json['title'],
+                source=post_json['source'],
+                origin=post_json['origin'],
+                description=post_json['description'],
+                author=author,
+                published=post_json['published'],
+            )
+            node_posts.append(post)
+
+    if node_posts == []:
+        node_posts = Post.objects.none()
+
+    return node_posts
